fix: correct protocol logic and remove unnecessary http2 dependency
- Restored usage of Go 1.24+ http.Transport.Protocols field for protocol configuration
- Removed redundant golang.org/x/net/http2 import
- Restored original API names: WithHTTPProxy, WithSocks5Proxy, and WithDNSResolver
- Ensured backward compatibility with the original implementation
- Verified all tests pass and project builds correctly
